requirements.txt
coverage==7.6.10
Django==5.1.4
django-picklefield==3.2
django-q==1.3.9
django-q2==1.7.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.1.0
requests==2.32.3
setuptools==75.7.0
stripe==11.4.1
